recommender_comparison/requirements.txt
pandas==2.2.2
numpy==1.25.2
scikit-learn==1.3.2
plotly==5.16.0
streamlit==1.24.0
scikit-surprise==1.1.3
requests==2.31.0
